+++ cropper/src/main/java/com/canhub/cropper/CropImageContractOptions.kt
@@ -490,6 +490,17 @@ data class CropImageContractOptions @JvmOverloads constructor(
         cropImageOptions.cropMenuCropButtonIcon = drawableResource
         return this
     }
+
+    /**
+     * Set whether the cropping option should be allowed or skipped entirely.<br></br>
+     * *Default: false*
+     */
+    fun setSkipEditing(skipEditing: Boolean): CropImageContractOptions {
+        cropImageOptions.skipEditing = skipEditing
+        cropImageOptions.showCropOverlay = !skipEditing
+        return this
+    }
+
     /**
      * Shows an intent chooser instead of the alert dialog when choosing an image source.
      * *Default: false*