+++ cropper/src/main/java/com/canhub/cropper/CropImageOptions.kt
@@ -285,6 +285,13 @@ open class CropImageOptions : Parcelable {
     var cropMenuCropButtonIcon: Int
 
     /**
+     * Allows you to skip the editing (cropping, flipping or rotating) option.
+     * This returns the entire selected image directly
+     */
+    @JvmField
+    var skipEditing: Boolean
+
+    /**
      * Enabling this option replaces the current AlertDialog to choose the image source
      * with an Intent chooser
      */
@@ -349,6 +356,7 @@ open class CropImageOptions : Parcelable {
         flipVertically = false
         cropMenuCropButtonTitle = null
         cropMenuCropButtonIcon = 0
+        skipEditing = false
         showIntentChooser = false
     }
 
@@ -408,6 +416,7 @@ open class CropImageOptions : Parcelable {
         flipVertically = parcel.readByte().toInt() != 0
         cropMenuCropButtonTitle = TextUtils.CHAR_SEQUENCE_CREATOR.createFromParcel(parcel)
         cropMenuCropButtonIcon = parcel.readInt()
+        skipEditing = parcel.readByte().toInt() != 0
         showIntentChooser = parcel.readByte().toInt() != 0
     }
 
@@ -466,6 +475,7 @@ open class CropImageOptions : Parcelable {
         dest.writeByte((if (flipVertically) 1 else 0).toByte())
         TextUtils.writeToParcel(cropMenuCropButtonTitle, dest, flags)
         dest.writeInt(cropMenuCropButtonIcon)
+        dest.writeByte((if (skipEditing) 1 else 0).toByte())
         dest.writeByte((if (showIntentChooser) 1 else 0).toByte())
     }